+++ Config/_bpp_functions.py
@@ -1,574 +1,470 @@
-import random, statistics, re
-from Config._const import ALPHABET
-from Config._functions import is_float, is_whole, strip_alpha, match_count
+import random, statistics, re, itertools, time, math
 import numpy as np
 
-def safe_multiply(a, b):
-	if isinstance(a, int) and isinstance(b, int) and \
-	    (abs(a) > 1<<512 or abs(b) > 1<<512):
-		raise ValueError('Numbers are too large to multiply')
-	return a*b
-		
-def safe_exponent(a, b):
-	if abs(a) > 1024 or b > 1024:
-		raise ValueError('Numbers are too large to exponentiate')
-	return a**b
+try:
+	from Config._const import ALPHABET
+	from Config._functions import is_float, is_whole, is_number, strip_alpha, match_count
+except ModuleNotFoundError:
+	from _const import ALPHABET
+	from _functions import is_float, is_whole, is_number, strip_alpha, match_count
 
-def safe_sort(a):
-	if len(a) > 1000:
-		raise ValueError('Iterable is too large to be sorted')
+class ProgramDefinedException(Exception): # used in THROW
+	pass
+
+def express_array(l):
+	str_form = " ".join(["\"" + str(a) + "\"" for a in l])
+	return f'[ARRAY {str_form}]'
+
+def safe_cut(s):
+	return str(s)[:15] + ("..." if len(str(s)) > 15 else "")
+
+def COMMENT(*a): return ""
+
+def INDEXOF(a,b,c=None,d=None):
+	if not is_number(c) and c is not None and not isinstance(c,str):
+		raise TypeError(f"Optional third parameter of INDEXOF function must be a number: {safe_cut(c)}")
+	if not is_number(d) and d is not None and not isinstance(d,str):
+		raise TypeError(f"Optional fourth parameter of INDEXOF function must be a number: {safe_cut(d)}")
+	if isinstance(c,str):
+		try:
+			c = int(c)
+		except:
+			raise TypeError(f"Optional third parameter of INDEXOF function must be a number: {safe_cut(c)}")
+
+	if isinstance(d,str):
+		try:
+			d = int(d)
+		except:
+			raise TypeError(f"Optional third parameter of INDEXOF function must be a number: {safe_cut(d)}")
+	if type(a) != str and not isinstance(a,list):
+		raise TypeError(f"First parameter of INDEXOF function must be an array or string: {safe_cut(a)}")
+	if c is not None:
+		if d is not None:
+			return a.index(b,c,d)
+		return a.index(b,c)
+	return a.index(b)
+
+def TIMEFUNC(): return time.time()
+
+def USERNAME(): return ("n", )
+
+def USERID(): return ("id", )
+
+def ABS(a):
+	if not is_number(a):
+		raise TypeError(f"Parameter of ABS function must be a number: {safe_cut(a)}")
+
+	return abs(int(a) if is_whole(a) else float(a))
+
+def SPLIT(a,b):
+	if type(a) == list:
+		raise TypeError(f"Parameter of SPLIT function cannot be an array: {safe_cut(a)}")
+	if type(b) == list:
+		raise TypeError(f"Parameter of SPLIT function cannot be an array: {safe_cut(b)}")
+	return str(a).split(str(b))
+
+def REPLACE(a,b,c):
+	if type(a) == list:
+		raise TypeError(f"Parameter of REPLACE function cannot be an array: {safe_cut(a)}")
+	return str(a).replace(str(b),str(c))
+
+def INDEX(a, b):
+	if type(a) not in [list, str]:
+		raise TypeError(f"First parameter of INDEX function must be a string or an array: {safe_cut(a)}")
+	if not is_whole(b):
+		raise TypeError(f"Second parameter of INDEX function must be an integer: {safe_cut(b)}")
+	return a[int(b)]
+
+def SLICE(a, b, c):
+	if not is_whole(b):
+		raise TypeError(f"Second parameter of SLICE function must be an integer: {safe_cut(b)}")
+	if not is_whole(c):
+		raise TypeError(f"Third parameter of SLICE function must be an integer: {safe_cut(c)}")
+	if type(a) == list:
+		to_cut = a
+	else:
+		to_cut = str(a)
+		
+	return to_cut[int(b):int(c)]
+
+def ARRAY(*a):
+	return list(a)
+
+def CONCAT(*a):
+	all_type = None
+	
+	for a1 in a:
+		if type(a1) in [int, float]: a1 = str(a1)
+		if all_type is None: all_type = type(a1)
+		elif type(a1) != all_type:
+			raise TypeError("CONCAT parameters must either be all arrays or all strings")
+	
+	if all_type == str:
+		a = [str(a1) for a1 in a]
+		return ''.join(a)
+	if all_type == list:
+		a = list(itertools.chain(*a))
+		return a
+	else:
+		raise IndexError("Cannot call CONCAT function with no arguments")
+
+def LENGTH(a):
+	if type(a) in [int, float]: a = str(a)
+	return len(a)
+
+def ARGS(a=None):
+	if not is_whole(a) and a is not None:
+		raise ValueError(f"ARGS function index must be an integer: {safe_cut(a)}")
+	
+	if a is None:
+		return ("aa", )
+	
+	return ("a", int(a))
+
+def GLOBALDEFINE(a, b):
+	if type(a) != str:
+		raise NameError(f"Global variable name must be a string: {safe_cut(a)}")
+
+	if re.search(r"[^A-Za-z_0-9]", a) or re.search(r"[0-9]", a[0]):
+		raise NameError(
+		f"Global variable name must be only letters, underscores and numbers, and cannot start with a number: {safe_cut(a)}")
+	
+	return ("gd", b)
+
+def GLOBALVAR(a):
+	if type(a) != str:
+		raise NameError(f"Global variable name must be a string: {safe_cut(a)}")
+
+	if re.search(r"[^A-Za-z_0-9]", a) or re.search(r"[0-9]", a[0]):
+		raise NameError(
+		f"Global variable name must be only letters, underscores and numbers, and cannot start with a number: {safe_cut(a)}")
+	
+	return ("gv", a)
+
+def DEFINE(a, b):
+	if type(a) != str:
+		raise NameError(f"Variable name must be a string: {safe_cut(a)}")
+
+	if re.search(r"[^A-Za-z_0-9]", a) or re.search(r"[0-9]", a[0]):
+		raise NameError(
+		f"Variable name must be only letters, underscores and numbers, and cannot start with a number: {safe_cut(a)}")
+	
+	return ("d", b)
+
+def VAR(a):
+	if type(a) != str:
+		raise NameError(f"Variable name must be a string: {safe_cut(a)}")
+
+	if re.search(r"[^A-Za-z_0-9]", a) or re.search(r"[0-9]", a[0]):
+		raise NameError(
+		f"Variable name must be only letters, underscores and numbers, and cannot start with a number: {safe_cut(a)}")
+	
+	return ("v", a)
+
+def REPEAT(a, b):
+	if not is_whole(b):
+		raise ValueError(f"Second parameter of REPEAT function is not an integer: {safe_cut(b)}")
+
+	if type(a) != list:
+		a = str(a)
+	b = int(b)
+
+	if b > 1024:
+		raise ValueError(f"Second parameter of REPEAT function is too large: {safe_cut(b)} (limit 1024)")
+
+	return a * b
+
+def CHOOSE(*a):
+	if len(a) == 1:
+		a = a[0]
+	return random.choice(a)
+
+def CHOOSECHAR(a):
+	if type(a) != str:
+		raise ValueError(f"CHOOSECHAR function parameter is not a string: {safe_cut(a)}")
+
+	return random.choice(list(a))
+
+def IF(a, b, c=""):
+	a = a not in [0, "0"]
+
+	if a:
+		return b
+	else:
+		return c
+
+def COMPARE(a, b, c):
+	operations = [">", "<", ">=", "<=", "!=", "=", "=="]
+	if b not in operations:
+		raise ValueError(f"Operation parameter of COMPARE function is not a comparison operator: {safe_cut(b)}")
+
+	if is_number(a): a = float(a)
+	if is_number(c): c = float(c)
+
+	if operations.index(b) <= 3 and type(a) != type(c):
+		raise TypeError(f"Entries to compare in COMPARE function are not the same type")
+
+	if b == ">": return int(a > c)
+	if b == "<": return int(a < c)
+	if b == ">=": return int(a >= c)
+	if b == "<=": return int(a <= c)
+	if b == "!=": return int(a != c)
+	if b == "=" or b == "==": return int(a == c)
+
+def MOD(a, b):
+	if not is_number(a):
+		raise ValueError(f"First parameter of MOD function is not a number: {safe_cut(a)}")
+	if not is_number(b):
+		raise ValueError(f"Second parameter of MOD function is not a number: {safe_cut(b)}")
+
+	a = int(a) if is_whole(a) else float(a)
+	b = int(b) if is_whole(b) else float(b)
+
+	if b == 0: raise ZeroDivisionError(f"Second parameter of MOD function cannot be zero")
+
+	return a % b
+
+def MATHFUNC(a, b, c):
+	operations = "+-*/^%"
+	if not is_number(a):
+		raise ValueError(f"First parameter of MATH function is not a number: {safe_cut(a)}")
+	if b not in operations:
+		raise ValueError(f"Operation parameter of MATH function not an operation: {safe_cut(b)}")
+	if not is_number(c):
+		raise ValueError(f"Second parameter of MATH function is not a number: {safe_cut(c)}")
+
+	a = int(a) if is_whole(a) else float(a)
+	c = int(c) if is_whole(c) else float(c)
+
+	if b == "+": return a+c
+	if b == "-": return a-c
+
+	if b == "*":
+		if abs(a) > 1e50:
+			raise ValueError(f"First parameter of MATH function too large to safely multiply: {safe_cut(a)} (limit 10^50)")
+		if abs(c) > 1e50:
+			raise ValueError(f"Second parameter of MATH function too large to safely multiply: {safe_cut(a)} (limit 10^50)")
+		return a*c
+
+	if b == "/":
+		if c == 0: raise ZeroDivisionError(f"Second parameter of MATH function in division cannot be zero")
+		return a/c
+	
+	if b == "%":
+		if c == 0: raise ZeroDivisionError(f"Second parameter of MATH function in modulo cannot be zero")
+		return a%c
+	
+	
+	if b == "^":
+		try:
+			return math.pow(a, c)
+		except OverflowError:
+			raise ValueError(f"Parameters of MATH function too large to safely exponentiate: {safe_cut(a)}, {safe_cut(c)}")
+
+def RANDINT(a, b):
+	if not is_whole(a):
+		raise ValueError(f"First parameter of RANDINT function is not an integer: {safe_cut(a)}")
+	if not is_whole(b):
+		raise ValueError(f"Second parameter of RANDINT function is not an integer: {safe_cut(b)}")
+
+	ints = [int(a), int(b)]
+	a, b = [min(ints), max(ints)]
+
+	if a == b: b += 1
+
+	return random.randrange(a, b)
+
+def RANDOM(a, b):
+	if not is_number(a):
+		raise ValueError(f"First parameter of RANDOM function is not a number: {safe_cut(a)}")
+	if not is_number(b):
+		raise ValueError(f"Second parameter of RANDOM function is not a number: {safe_cut(b)}")
+
+	a = float(a)
+	b = float(b)
+
+	return random.uniform(a, b)
+
+def THROW(a): # don't need to check, because either way it'll error
+	raise ProgramDefinedException(a)
+	
+def TYPEFUNC(a):
+	if is_whole(a): return "int"
+	if is_number(a): return "float"
+	return type(a).__name__
+	
+def ROUND(a, b=0):
+	if not is_number(a):
+		raise ValueError(f"ROUND function parameter is not a number: {safe_cut(a)}")
+	if not is_whole(b):
+		raise ValueError(f"ROUND function parameter is not an integer: {safe_cut(b)}")
+	
+	rounded = round(float(a), int(b))
+	if rounded.is_integer(): return int(rounded)
+	return rounded
+
+def FLOOR(a):
+	if not is_number(a):
+		raise ValueError(f"FLOOR function parameter is not a number: {safe_cut(a)}")
+	
+	return math.floor(float(a))
+
+def CEIL(a):
+	if not is_number(a):
+		raise ValueError(f"CEIL function parameter is not a number: {safe_cut(a)}")
+	
+	return math.ceil(float(a))
+
+def LOG(a, b):
+	if not is_number(a):
+		raise ValueError(f"LOG function parameter is not a number: {safe_cut(a)}")
+	if not is_number(b):
+		raise ValueError(f"LOG function parameter is not a number: {safe_cut(b)}")
+		
+	if b == 0: raise ValueError("Second parameter of LOG function must not be zero")
+		
+	return math.log(float(a),float(b))
+
+def FACTORIAL(a):
+	if not is_number(a):
+		raise ValueError(f"FACTORIAL function parameter is not a number: {safe_cut(a)}")
+	
+	try:
+		return math.gamma(float(a)+1) # extension of the factorial function, allows floats too
+	except OverflowError:
+		raise ValueError(f"First parameter of FACTORIAL function too large to safely factorial: {safe_cut(a)}")
+
+def SIN(a):
+	if not is_number(a):
+		raise ValueError(f"SIN function parameter is not a number: {safe_cut(a)}")
+	
+	return math.sin(float(a))
+
+def TAN(a):
+	if not is_number(a):
+		raise ValueError(f"TAN function parameter is not a number: {safe_cut(a)}")
+	
+	return math.tan(float(a))
+
+def COS(a):
+	if not is_number(a):
+		raise ValueError(f"COS function parameter is not a number: {safe_cut(a)}")
+	
+	return math.cos(float(a))
+
+def MINFUNC(a):
+	if not type(a) == list:
+		raise ValueError(f"MIN function parameter is not a list: {safe_cut(a)}")
+	
+	if 0 not in [is_number(elem) for elem in a]:
+		a = [float(elem) for elem in a]
+	
+	return min(a)
+
+def MAXFUNC(a):
+	if not type(a) == list:
+		raise ValueError(f"MAX function parameter is not a list: {safe_cut(a)}")
+	
+	if 0 not in [is_number(elem) for elem in a]:
+		a = [float(elem) for elem in a]
+	
+	return max(a)
+
+def SHUFFLE(a):
+	if not type(a) == list:
+		raise ValueError(f"SHUFFLE function parameter is not a list: {safe_cut(a)}")
+		
+	return random.sample(a, k=len(a))
+
+def SORTFUNC(a):
+	if not type(a) == list:
+		raise ValueError(f"SORT function parameter is not a list: {safe_cut(a)}")
+	
+	if 0 not in [is_number(elem) for elem in a]:
+		a = [float(elem) for elem in a]
+	
 	return sorted(a)
 
-def safe_desort(a):
-	if len(a) > 1000:
-		raise ValueError('Iterable is too large to be sorted')
-	return sorted(a, reverse=True)
+def JOIN(a, b=""):
+	if not type(a) == list:
+		raise ValueError(f"First JOIN function parameter is not a list: {safe_cut(a)}")
+	if not type(b) == str:
+		raise ValueError(f"Second JOIN function parameter is not a string: {safe_cut(b)}")
+	
+	a = [str(elem) for elem in a]
+	
+	return b.join(a)
+	
+def SETINDEX(a, b, c):
+	if not is_whole(b):
+		raise ValueError(f"SETINDEX function parameter is not an integer: {safe_cut(b)}")
+	if type(a) == list:
+		mylist = a.copy()
+		mylist[int(b)] = c
+		return mylist
+	a = str(a)
+	if len(str(c)) > 1:
+		raise ValueError(f"SETINDEX function paramater is not a character: {safe_cut(c)}")
+	return a[0:int(b)] + c + a[int(b)+1:]
 
-def safe_concat(a, b):
-	if len(a + b) > 500:
-		raise ValueError('Strings are too large to be concatenated')
-	return a + b
+def UNICODE(a):
+	if len(str(a)) != 1:
+		ValueError(f"CHAR function paramater is not a character: {safe_cut(a)}")
+	return ord(str(a))
+
+def CHARFUNC(a):
+	if not is_whole(a):
+		raise ValueError(f"CHAR function parameter is not an integer: {safe_cut(a)}")
+	try:
+		out = chr(a)
+	except:
+		raise ValueError(f"CHAR function parameter is not a valid character: {safe_cut(a)}")
+	return out
 
 FUNCTIONS = {
-
-	# Basic operations
-	"out{?}": {
-		"TYPES": {
-			"a": []
-		}
-	},
-
-
-	# Mathematical
-	"?\+?": {
-		"MAIN": (lambda a, b: a + b),
-		"TYPES": {
-			"a": ["NUMBER"],
-			"b": ["NUMBER"]
-		},
-		"PRIORITY": 1
-	},
-
-	"?-?": {
-		"MAIN": (lambda a, b: a - b),
-		"TYPES": {
-			"a": ["NUMBER"],
-			"b": ["NUMBER"]
-		},
-		"PRIORITY": 1
-	},
-
-	"?\*?": {
-		"MAIN": safe_multiply,
-		"TYPES": {
-			"a": ["NUMBER"],
-			"b": ["NUMBER"]
-		},
-		"PRIORITY": 2
-	},
-
-	"?/?": {
-		"MAIN": (lambda a, b: a / b),
-		"TYPES": {
-			"a": ["NUMBER"],
-			"b": ["NUMBER"]
-		},
-		"PRIORITY": 2
-	},
-
-	"?\^?": {
-		"MAIN": safe_exponent,
-		"TYPES": {
-			"a": ["NUMBER"],
-			"b": ["NUMBER"]
-		},
-		"PRIORITY": 3
-	},
-
-	"avg?": {
-		"MAIN": (lambda a: np.mean(a)),
-		"TYPES": {
-			"a": ["ARRAY NUMBER"]
-		}
-	},
-
-	"stdev?": {
-		"MAIN": (lambda a: statistics.stdev(a)),
-		"TYPES": {
-			"a": ["ARRAY NUMBER"]
-		}
-	},
-
-	"stdevp?": {
-		"MAIN": (lambda a: np.std(a)),
-		"TYPES": {
-			"a": ["ARRAY NUMBER"]
-		}
-	},
-
-	"randint?to?": {
-		"MAIN": (lambda a, b: random.randrange(a, b+1)),
-		"TYPES": {
-			"a": ["INTEGER"],
-			"b": ["INTEGER"]
-		}
-	},
-
-	"rand?to?": {
-		"MAIN": (lambda a, b: random.uniform(a, b)),
-		"TYPES": {
-			"a": ["NUMBER"],
-			"b": ["NUMBER"]
-		}
-	},
-
-
-	# STRING and ARRAY manipulation
-	"?index?": {
-		"MAIN": (lambda a, b: a[b-1] if b >= 0 else a[b]),
-		"TYPES": {
-			"a": ["ARRAY", "STRING"],
-			"b": ["INTEGER"]
-		},
-		"PRIORITY": 4
-	},
-
-	"?slice?to?": {
-		"MAIN": (lambda a, b, c: a[b-1:c-1]),
-		"TYPES": {
-			"a": ["ARRAY", "STRING"],
-			"b": ["INTEGER"],
-			"c": ["INTEGER"]
-		},
-		"PRIORITY": 4
-	},
-
-	"?&?": {
-		"MAIN": safe_concat,
-		"TYPES": {
-			"a": ["ARRAY", "STRING"],
-			"b": ["ARRAY", "STRING"]
-		},
-		"PRIORITY": 4
-	},
-
-	"?rept?": {
-		"MAIN": (lambda a, b: a * b),
-		"TYPES": {
-			"a": ["ARRAY", "STRING"],
-			"b": ["INTEGER"]
-		}
-	},
-
-	"?choose?": {
-		"MAIN": (lambda a, b: (random.sample(a, b) if b != 1 else random.sample(a, b)[0])),
-		"TYPES": {
-			"a": ["ARRAY"],
-			"b": ["INTEGER"]
-		}
-	},
-
-	"sort?": {
-		"MAIN": safe_sort,
-		"TYPES": {
-			"a": ["ARRAY NUMBER", "ARRAY STRING", "ARRAY BOOLEAN"]
-		}
-	},
-
-	"desort?": {
-		"MAIN": safe_desort,
-		"TYPES": {
-			"a": ["ARRAY NUMBER", "ARRAY STRING", "ARRAY BOOLEAN"]
-		}
-	},
-
-
-	# Conditional
-	"?==?": {
-		"MAIN": (lambda a, b: a == b),
-		"TYPES": {
-			"a": [],
-			"b": []
-		},
-		"PRIORITY": 0
-	},
-
-	"?>?": {
-		"MAIN": (lambda a, b: a > b),
-		"TYPES": {
-			"a": [],
-			"b": []
-		},
-		"PRIORITY": 0
-	},
-
-	"?<?": {
-		"MAIN": (lambda a, b: a < b),
-		"TYPES": {
-			"a": [],
-			"b": []
-		},
-		"PRIORITY": 0
-	},
-
-	"?>=?": {
-		"MAIN": (lambda a, b: a >= b),
-		"TYPES": {
-			"a": [],
-			"b": []
-		},
-		"PRIORITY": 0
-	},
-
-	"?<=?": {
-		"MAIN": (lambda a, b: a <= b),
-		"TYPES": {
-			"a": [],
-			"b": []
-		},
-		"PRIORITY": 0
-	},
-
-	"?!=?": {
-		"MAIN": (lambda a, b: a != b),
-		"TYPES": {
-			"a": [],
-			"b": []
-		},
-		"PRIORITY": 0
-	},
-
-	"!?": {
-		"MAIN": (lambda a: not a),
-		"TYPES": {
-			"a": ["BOOLEAN"]
-		},
-		"PRIORITY": 1
-	},
+	"MATH": MATHFUNC,
+	"RANDINT": RANDINT,
+	"RANDOM": RANDOM,
+	"FLOOR": FLOOR,
+	"CEIL": CEIL,
+	"ROUND": ROUND,
+	"COMPARE": COMPARE,
+	"IF": IF,
+	"CHOOSE": CHOOSE,
+	"CHOOSECHAR": CHOOSECHAR,
+	"REPEAT": REPEAT,
+	"DEFINE": DEFINE,
+	"VAR": VAR,
+	"CONCAT": CONCAT,
+	"ARRAY": ARRAY,
+	"INDEX": INDEX,
+	"INDEXOF": INDEXOF,
+	"ARGS": ARGS,
+	"ABS": ABS,
+	"#": COMMENT,
+	"GLOBAL DEFINE": GLOBALDEFINE,
+	"GLOBAL VAR": GLOBALVAR,
+	"MOD": MOD,
+	"LENGTH": LENGTH,
+	"USERNAME": USERNAME,
+	"USERID": USERID,
+	"SLICE": SLICE,
+	"REPLACE": REPLACE,
+	"SPLIT": SPLIT,
+	"TIME": TIMEFUNC,
+	"LOG": LOG,
+	"FACTORIAL": FACTORIAL,
+	"SIN": SIN,
+	"COS": COS,
+	"TAN": TAN,
+	"TYPE": TYPEFUNC,
+	"THROW": THROW,
+	"MIN": MINFUNC,
+	"MAX": MAXFUNC,
+	"SHUFFLE": SHUFFLE,
+	"SORT": SORTFUNC,
+	"JOIN": JOIN,
+	"SETINDEX": SETINDEX,
+	"CHAR" : CHARFUNC,
+	"UNICODE" : UNICODE
 }
-
-# Tries to interpret the code passed as an operation
-# Return format: [is_it_an_operation, result_of_operation, was_there_an_error]
-def operation_check(block):
-	matching_ops = [] # List of operations that match
-
-	for possible_op in FUNCTIONS.keys(): # Check through each operation
-		letter_expect = False
-		op_regex = possible_op # Prepare a variable to serve as the regex for whether or not the code matches 
-		# this operation
-
-		for param in range(possible_op.count("?")): # Each ? is a parameter of the function
-			alph = ALPHABET.lower()[param]
-			types = FUNCTIONS[possible_op]["TYPES"][alph] # Find out what type is expected for that parameter
-
-			if "STRING" in types or "BOOLEAN" in types or "ARRAY STRING" in types or "ARRAY BOOLEAN" in types:
-				letter_expect = True
-			
-			dot_exp = r"([^\^!-&\*-\/<-@_]{1,})"
-			if possible_op == "out{?}":
-				dot_exp = r"([^\^]{1,})"
-			elif types == ["NUMBER"] or types == ["INTEGER"] or types == ["ARRAY NUMBER"]:
-				dot_exp = r"([^\\A-Za-z]{1,})" # A-z would include square brackets, which are necessary in ARRAY NUMBERs
-			'''if len(strip_alpha(possible_op)) == 0:
-				dot_exp = r"([^\^\\!-\/<-@]{1,})"
-			elif len(types) == 0:
-				dot_exp = r"([^\\]{1,})"
-			elif types == ["NUMBER"] or types == ["INTEGER"] or types == ["ARRAY NUMBER"]:
-				# If it's expecting a number of some kind, the parameter shouldn't have any letters in it, and so we 
-				# can use a regex that does not count letter characters to detect this parameter
-				dot_exp = r"([^A-z]{1,})"'''
-			
-			# If any type is allowed (len(types) == 0) or the type is BOOLEAN or STRING (which contain letters)
-			# Use the regex that matches all characters except for backslashes
-
-			op_regex = op_regex.replace("?", dot_exp, 1) # Replace the _ placeholder with the matching group for that
-			# parameter
-		
-		match = re.search(op_regex, block) # Try to match this operation (the previously generated regex) with the code
-		
-		if match: # If it's a match, append it to the matching operations array
-			if possible_op.count("?") == 2 and "\\" == match.group(1)[-1]:
-				continue
-			
-			if not possible_op.startswith("?") and match.span()[0] != 0 and block[match.span()[0]-1] == "\\":
-				continue
-			
-			matching_ops.append([possible_op, match, letter_expect])
-
-
-	if "out{?}" in matching_ops or "out{" in block: # If the operation is out{}, leave a flag for it. It'll be handled
-		substring = block[block.find("out{")+4:block.find("}")]
-		return ["out", f"({substring})", False] # in parenthesis_parser
-
-	if len(matching_ops) == 0: # If no operations match the code, return it unchanged
-		return [False, block, False]
-	
-	letter_operations = [x for x in matching_ops if strip_alpha(x[0]) != ""] # Operations that don't use letters are
-
-	if len(letter_operations) == 0 and strip_alpha(block) != "":
-		if True not in [x[2] for x in matching_ops]:
-			return [False, block, False]
-		
-	if len(letter_operations) == 0 and len(matching_ops) != 0: # all either math operations or other operations that
-	# can be used with eachother. If this statement is composed entirely of them, then we can treat it differently
-
-		operators = [
-			[x[0].replace("?", "").replace("\\", ""), 
-			FUNCTIONS[x[0]]["PRIORITY"],
-			x[0]]
-			for x in matching_ops
-		]
-
-		full_token_list = []
-		current_index = 0
-		omit_bracket = False
-
-		while True:
-			operator_indices = []
-			for op in operators:
-				found = block.find(op[0], current_index)
-				if found != -1:
-					operator_indices.append(found)
-				else:
-					operator_indices.append(len(block))
-
-			min_op = min(operator_indices)
-			now_op = operators[operator_indices.index(min_op)][0]
-
-			if min_op == len(block):
-				full_token_list.append(block[current_index:])
-				break
-
-			to_add = block[current_index:min_op]
-
-			full_token_list.append(to_add)
-			full_token_list.append(now_op)
-			current_index = min_op + len(now_op)
-		
-		#full_token_list.append(block[current_index:])
-		
-		#operator_order = sorted(operator_order, key=lambda o: (5 - o[1]))
-
-
-		priority_limit = max([x[1] for x in operators])
-		op_list = [x[0] for x in operators]
-
-		for p in range(priority_limit + 1):
-			current_priority = priority_limit - p
-
-			for token in full_token_list:
-				if token in op_list:
-					ind = op_list.index(token)
-					if operators[ind][1] != current_priority:
-						continue
-					
-					token_ind = full_token_list.index(token)
-					operation_name = operators[ind][2]
-					param_info = FUNCTIONS[operation_name]["TYPES"]
-					
-					params = []
-					params_ind = []
-					if len(param_info.keys()) != 1:
-						for r in range(token_ind):
-							search_ind = token_ind - r - 1
-							if full_token_list[search_ind] != []:
-								c_param = full_token_list[search_ind]
-
-								param_name = "a"
-								expected_types = param_info[param_name] # The allowed types for param
-								fit_type = "" # The type this parameter will fit under
-
-								for expect in expected_types:
-									if expect == "NUMBER":
-										if is_whole(c_param):
-											c_param = int(c_param)
-											fit_type = expect
-											break
-										if is_float(c_param):
-											c_param = float(c_param)
-											fit_type = expect
-											break
-									if expect == "INTEGER":
-										if is_whole(c_param):
-											c_param = int(c_param)
-											fit_type = expect
-											break
-									if expect == "BOOLEAN":
-										if c_param in ["True", "False"]:
-											c_param = bool(c_param)
-											fit_type = expect
-											break
-									if expect.startswith("ARRAY"):
-										if c_param.startswith("[") and c_param.endswith("]"):
-											c_param = array_to_list(c_param)
-											fit_type = expect
-											break
-									if expect == "STRING":
-										if (not is_float(c_param) and not is_whole(c_param)
-										and c_param not in ["True", "False"]):
-											fit_type = expect
-											break
-								
-								if fit_type == "" and len(expected_types) > 0:
-									#return [True, [operation_name, param_name, c_param], True]
-									return [True, block, True]
-								
-								params_ind.append(search_ind)
-								params.append(c_param)
-								break
-
-					for search_ind in range(token_ind + 1, len(full_token_list)):
-						if full_token_list[search_ind] != []:
-							c_param = full_token_list[search_ind]
-
-							param_name = "a" if len(param_info.keys()) == 1 else "b"
-							expected_types = param_info[param_name]
-							fit_type = "" # The type this parameter will fit under
-
-							for expect in expected_types:
-								if expect == "NUMBER":
-									if is_whole(c_param):
-										c_param = int(c_param)
-										fit_type = expect
-										break
-									if is_float(c_param):
-										c_param = float(c_param)
-										fit_type = expect
-										break
-								if expect == "INTEGER":
-									if is_whole(c_param):
-										c_param = int(c_param)
-										fit_type = expect
-										break
-								if expect == "BOOLEAN":
-									if c_param in ["True", "False"]:
-										c_param = bool(c_param)
-										fit_type = expect
-										break
-								if expect.startswith("ARRAY"):
-									if c_param.startswith("[") and c_param.endswith("]"):
-										c_param = array_to_list(c_param)
-										fit_type = expect
-										break
-								if expect == "STRING":
-									if (not is_float(c_param) and not is_whole(c_param)
-									and c_param not in ["True", "False"]):
-										fit_type = expect
-										break
-							
-							if fit_type == "" and len(expected_types) > 0:
-								#return [True, [operation_name, param_name, c_param], True]
-								return [True, block, True]
-							
-							params_ind.append(search_ind)
-							params.append(c_param)
-							break
-					
-					result = FUNCTIONS[operation_name]["MAIN"](*params)
-
-					if type(result) is list:
-						result = list_to_array(result)
-					
-					full_token_list[token_ind] = str(result)
-					for x in params_ind:
-						full_token_list[x] = []
-			
-			full_token_list = [x for x in full_token_list if x != []]
-		
-		full_token_list = [x for x in full_token_list if x.strip() != ""]
-
-		return [True, full_token_list[0], False] # There was an operation and no error. Pass the result
-
-	# Just in case there's more than 1 matching operation, run this sorted command. It takes the span of the operation
-	# match (the start and end indices for when the operation applies), and subtracts the start index from end index 
-	# to determine how long the span is. Put the one with the longest span first
-	matching_ops = sorted(matching_ops, reverse=True, key=(lambda m: len(m[0])))
-	matching_ops = sorted(matching_ops, reverse=True, key=(lambda m: m[1].span()[1] - m[1].span()[0]))
-
-	# Assume it's the one with the longest span. This prevents subsets of operations being counted instead of the
-	operation, match, lt = matching_ops[0] # actual operation specified. Define the operation and the match object
-	
-	parameters = [] # List of parameters that will be used in the operation function
-
-	for g in range(operation.count("?")): # For each parameter necessary...
-		param = match.group(g + 1).strip() # Match the group of the parameter (group 0 is the entire block; start at 1)
-		param_name = ALPHABET.lower()[g] # Get the parameter name (which is just the corresponding letter in alphabet)
-		
-		expected_types = FUNCTIONS[operation]["TYPES"][param_name] # The allowed types for this parameter
-		fit_type = "" # The type this parameter will fit under
-
-		for expect in expected_types:
-			if expect == "NUMBER": # NUMBER can be both int and float
-				if is_whole(param): # Detect if it can be interpreted as an int
-					param = int(param) # If so, make it an int
-					fit_type = expect
-					break
-				if is_float(param): # Detect if it can be interpreted as a float
-					param = float(param) # If so, make it a float
-					fit_type = expect
-					break
-			if expect == "INTEGER": # INTEGER is exclusively int
-				if is_whole(param): # Detect if it can be an int
-					param = int(param) # If so, make it an int
-					fit_type = expect
-					break
-			if expect == "BOOLEAN": # BOOLEAN can only be True or False
-				if param in ["True", "False"]: # Detect if it's one of those
-					param = bool(param) # If so, turn it into a bool
-					fit_type = expect
-					break
-			if expect.startswith("ARRAY"): # ARRAYs are defined with brackets
-				if param.startswith("[") and param.endswith("]"):
-					param = array_to_list(param)
-					fit_type = expect
-					break
-			if expect == "STRING": # STRING can be anything that isn't instantly defined as one of the others
-				if not is_float(param) and not is_whole(param) and param not in ["True", "False"]: # Detect that it's
-					# not any of the other types. param is already a str, so it doesn't need to be made into one
-					fit_type = expect
-					break
-
-		if fit_type == "" and len(expected_types) > 0: # If there ARE expected types but the parameter can't fit them
-			#return [True, [operation_name, param_name, c_param], True]
-			return [True, block, True] # There IS an operation but there is an error.
-			# Return the operation and param name for error specification
-			
-		parameters.append(param) # If everything went alright, add param to the parameter list
-
-	result = FUNCTIONS[operation]["MAIN"](*parameters) # Extract a result from the operation function
-
-	return [True, result, False] # There was an operation and no error. Pass the result
-
-
-# Tries to interpret the code passed as a variable call
-# Return format: [is_it_a_variable, result]
-def variables(block, VARIABLES):
-	try: # Try to get the variable from the VARIABLES dict
-		result = VARIABLES[block]
-		return [True, result]
-	except KeyError: # If you can't, it's not a valid variable
-		return [False, block]
-
-
-# Returns a python list with the input as a B++ array
-def array_to_list(raw):
-	if "\t" not in raw:
-		list_eq = raw[1:-1].split("; ")
-	else:
-		list_eq = raw[1:-1].split("\t ")
-
-	for entry in range(len(list_eq)):
-		if is_whole(list_eq[entry]):
-			list_eq[entry] = int(list_eq[entry])
-		if is_float(list_eq[entry]):
-			list_eq[entry] = float(list_eq[entry])
-		if list_eq[entry] in ["True", "False"]:
-			list_eq[entry] = bool(list_eq[entry])
-
-	return list_eq
-
-
-# Returns a B++ array with input as a python list
-def list_to_array(formatted):
-	array_eq = f"[{'; '.join([str(x) for x in formatted])}]"
-	return array_eq
