+++ Config/b_star_interpreter/Codebase.py
@@ -30,3 +30,5 @@ class Codebase:
         # TODO: This is temporary while s3.py gets revamped in 1.1
         self.global_limit: int = 0
         self.ret = None
+
+
