+++ Config/b_star_interpreter/functions/slice.py
@@ -1,6 +1,8 @@
 
 
 def slice_func(array, index_start, index_end, index_step):
+
+
     if index_step is None:
         index_step = 1
     else:
